3.1.12 Sprawdzono projekt pod kątem klas które nie muszą być publiczne:
