+++ src/main/java/edu/kis/vh/nursery/FIFORhymer.java
@@ -1,5 +1,9 @@
 package edu.kis.vh.nursery;
 
+/**
+ * Klasa FIFORhymer została publiczna ponieważ inaczej fabryka nie miała by do niej dostępu
+ */
+
 public class FIFORhymer extends DefaultCountingOutRhymer {
 
 	public DefaultCountingOutRhymer counitngOutRhymer = new DefaultCountingOutRhymer();
